+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## in development:
 * **New Features:**
     * new primitive in "looks": NEW COSTUME from a list of pixels and dimensions, allowing CURRENT
+    * new primitige in "sound": NEW SOUND from a list of samples
     * added selectors for sprites' and the stage's bounding box (LEFT, RIGHT, TOP, BOTTOM) to MY dropdown
 * **Notable Changes:**
     * running STOP ALL now also toggles (pauses and resumes) all generic WHEN hat blocks (just like pressing the stop button)
@@ -11,6 +12,9 @@
     * assert that dimensions given for STRETCH are finite numbers (avoid crash)
 * **Translation Updates:**
     * German
+
+### 2019-10-20
+* objects, threads: added "new sound" from list of samples primitive reporter to "sound" category
 
 ### 2019-10-18
 * objects, blocks, threads: added dimension getters for the stage